+++ test_backtest/backtest.py
@@ -92,7 +92,5 @@ def run_minbacktest():
 
 if __name__ == '__main__':
     run_daybacktest()
-    run_minbacktest()
-    # backtest._settle()
+    #run_minbacktest()
 
-    # backtest.run()